[@mantine/form] Remove old use-form hook
