--- src/mantine-form/src/get-error-path/get-error-path.ts
@@ -1,3 +1,0 @@
-export function getErrorPath(path: (string | number | symbol)[]) {
-  return path.join('.');
-}
