+++ src/mantine-form/src/index.ts
@@ -1,8 +1,6 @@
-export { formList, isFormList } from './form-list/form-list';
 export { useForm } from './use-form';
 export { zodResolver } from './resolvers/zod-resolver/zod-resolver';
 export { yupResolver } from './resolvers/yup-resolver/yup-resolver';
 export { joiResolver } from './resolvers/joi-resolver/joi-resolver';
 
-export type { FormErrors, FormRules } from './types';
-export type { FormList } from './form-list/form-list';
+export type { FormErrors } from './types';
