+++ src/mantine-form/src/tests/clearErrors.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/clearErrors', () => {
   it('clears errors when clearErrors handler is called', () => {
