+++ src/mantine-form/src/tests/clearFieldError.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/clearFieldError', () => {
   it('clears error of given field with clearFieldError handler', () => {
