+++ src/mantine-form/src/tests/get-input-props.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/get-input-props', () => {
   it('returns correct input props (root property)', () => {
