+++ src/mantine-form/src/tests/setErrors.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/setErrors', () => {
   it('filters out errors with undefined and null with setErrors handler', () => {
