+++ src/mantine-form/src/tests/validate-nested.test.ts
@@ -1,5 +1,5 @@
 import { act, renderHook } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/validate with nested rules', () => {
   it('validates object with nested values', () => {
