+++ src/mantine-form/src/tests/validate-object.test.ts
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/validate with record rules', () => {
   it('validates all fields with validate handler', () => {
