+++ src/mantine-form/src/tests/values.test.ts
@@ -1,5 +1,5 @@
 import { renderHook, act } from '@testing-library/react';
-import { useForm } from '../use-form2';
+import { useForm } from '../use-form';
 
 describe('@mantine/form/values', () => {
   it('sets given initial values', () => {
